requirements.txt
paho-mqtt==1.6.1
Flask==2.3.3
Flask-SocketIO==5.3.5
eventlet==0.33.3
supabase==2.4.0
python-dotenv==1.0.0 